Alter some styles related to blockquote brutally
The left margin of blockquotes is shortened to reduce line-wraps when
content is viewed on a narrow screen, and the color of text in blockquotes
is return to black to make content more readable.
These changes are specified to my own site since content are largely wrapped
inside blockquotes, and nested blockquotes are commonly used there.